Question 1/requirements.txt
psutil
mysql-connector-python
multiprocessing
cryptography
